add usecases

+++ knowledgebase/commands/domain/models/knowledge_item.go
@@ -5,7 +5,7 @@ import "time"
 
 // KnowledgeItem represents one particular piece of knowledge.
 type KnowledgeItem struct {
-	ID     int    `json:"id"`
+	ID     int64  `json:"id"`
 	Title  string `json:"title"`
 	Anchor string `json:"anchor"`
 	Data   string `json:"description"`
@@ -14,9 +14,9 @@ type KnowledgeItem struct {
 
 	Tags []string `json:"tags,omitempty"`
 
-	Score int `json:"score"`
+	Score int64 `json:"score"`
 
-	LastMark    int        `json:"last_mark"`
+	LastMark    int64      `json:"last_mark"`
 	LastCheckAt *time.Time `json:"last_check_at"`
 
 	CreatedAt *time.Time `json:"created_at"`
